+++ ToyProject/app/src/main/java/com/egiwon/moviesearch/ui/MainActivity.kt
@@ -2,6 +2,9 @@ package com.egiwon.moviesearch.ui
 
 import android.content.Intent
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
+import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.app.ActivityOptionsCompat
 import androidx.lifecycle.Observer
 import com.egiwon.moviesearch.BR
@@ -40,6 +43,34 @@ class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>(R.layout.a
             }
         ) {}
         rvMovieList.setHasFixedSize(true)
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.main_menu, menu)
+
+        val nightMode = AppCompatDelegate.getDefaultNightMode()
+        if (nightMode == AppCompatDelegate.MODE_NIGHT_YES) {
+            menu?.findItem(R.id.night_mode)?.setIcon(R.drawable.ic_sunny)
+        } else {
+            menu?.findItem(R.id.night_mode)?.setIcon(R.drawable.ic_brightness)
+        }
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.night_mode -> {
+                val nightMode = AppCompatDelegate.getDefaultNightMode()
+                if (nightMode == AppCompatDelegate.MODE_NIGHT_YES) {
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+                } else {
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+                }
+                recreate()
+            }
+
+        }
+        return true
     }
 
     private fun observingViewModel() {
